Adding in support for multiple SALTs thus allowing them to be rotated
This commit adds in support for the SALT to be rotated by defining a list of salts within settings.py where the newer salts are added to the start.  

The first key will be used to encrypt all new data, and decryption of existing values will be attempted with all given keys in order. This is useful for key rotation: place a new key at the head of the list for use with all new or changed data, but existing values encrypted with old keys will still be accessible

This is based of django-fernet-fields which is a dead package but has some useful features such as this to allow the salt to be rotated in the future for a stronger salt.

```
SALT_KEY = [
    'my-newer-salt',
    'the-original-salt'
]
```